+++ TDTChocolate/TDTFoundationAdditions.h
@@ -27,3 +27,4 @@
 #import "FoundationAdditions/NSArray+TDTNullNormalization.h"
 #import "FoundationAdditions/NSDictionary+TDTNullNormalization.h"
 #import "FoundationAdditions/NSDictionary+TDTMerging.h"
+#import "FoundationAdditions/NSCalendar+TDTComponents.h"